Fixes #446 - improve test coverage
* Added direct tests from some new classes
* Broke up some of the geocoding tests to not be so huge
