+++ gedbrowser-renderer/src/main/java/org/schoellerfamily/gedbrowser/renderer/SourceRenderer.java
@@ -13,8 +13,9 @@ import org.schoellerfamily.gedbrowser.datamodel.visitor.SourceVisitor;
  * @author Dick Schoeller
  */
 public final class SourceRenderer extends GedRenderer<Source>
-        implements IndexHrefRenderer<Source>, HeaderHrefRenderer<Source>,
-        SubmittorsHrefRenderer<Source>, SourcesHrefRenderer<Source> {
+        implements HeaderHrefRenderer<Source>, IndexHrefRenderer<Source>,
+            PlacesHrefRenderer<Source>, SourcesHrefRenderer<Source>,
+            SubmittorsHrefRenderer<Source> {
     /**
      * @param gedObject the Source that we are going to render
      * @param rendererFactory the factory that creates the renderers for the
