+++ gedbrowser-renderer/src/test/java/org/schoellerfamily/gedbrowser/renderer/test/LivingRendererTest.java
@@ -222,4 +222,17 @@ public final class LivingRendererTest {
         assertEquals("sources href mismatch",
                 "submittors?db=gl120368", renderer.getSubmittorsHref());
     }
+
+    /**
+     * Test whether the menu items are as expected.
+     *
+     * @throws IOException if can't read data file
+     */
+    @Test
+    public void testPlacesMenuItem() throws IOException {
+        final Root root1 = reader.readFileTestSource();
+        final LivingRenderer renderer = new LivingRenderer(root1, userContext);
+        assertEquals("places href mismatch", "places?db=gl120368",
+                renderer.getPlacesHref());
+    }
 }
