+++ geoservice-persistence/src/test/java/org/schoellerfamily/geoservice/model/test/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Copyright 2016-2017 Richard Schoeller
+ * Tests for data model for the public visibility of geocoded data.
+ */
+package org.schoellerfamily.geoservice.model.test;
